requirements.txt
aiocron~=1.8
aiohttp~=3.8.4
black~=23.1.0
croniter~=1.3.8
flake8~=6.0.0
pytest~=7.2.2
python-dotenv~=1.0.0
slack-bolt~=1.16.4
slack-sdk~=3.20.2
typing~=3.7.4.3
